feat: add multiple pages with router and zustand
