+++ src/pages/Home.jsx
@@ -1,0 +1,8 @@
+const Home = () => (
+  <div>
+    <h2>欢迎来到示例应用</h2>
+    <p>使用上方导航浏览不同示例页面。</p>
+  </div>
+)
+
+export default Home
